requirements.txt
requests>=2.31.0
schedule>=1.2.0
scikit-learn>=1.3.0
numpy>=1.24.0
lightgbm>=4.0.0
nba_api>=1.4.0
python-dotenv>=1.0.0
